--- src/components/UI/inputs/index.ts
@@ -1,5 +1,0 @@
-export {default as BooleanInput} from "Boolean.svelte" 
-export {default as Button} from "Button.svelte"
-export {default as Textfield} from "TextField.svelte"
-
-// export default {BooleanInput, Button, Textfield}
