--- src/components/UI/tabs/Tab.svelte
@@ -1,20 +1,0 @@
-<script lang="ts">
-    export let tab: string;
-    let _class: string = "";
-    export {_class as class}
-</script>
-
-<template lang="pug">
-    div(class!="tab {_class}")
-        slot(item!="{tab}") {tab}
-</template>
-
-<style lang="scss">
-    .tab{
-        text-align: center;
-        text-justify: center;
-    }
-    .tab:hover{
-        cursor: pointer;
-    }
-</style>
